Adding logic to mantain the user session with a custom hook
